backend/requirements/prod.txt
django==5.0.3
python-dotenv
psycopg2
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
django-cors-headers
celery
django-rest-passwordreset
redis
django-celery-beat